fix upgradeReq is undefined
an optimization in version 3.0 of the required proxy server [ https://github.com/websockets/ws/issues/1114 ]drops the upgradeReq parameter to save 20% in memory.  This breaks non-http(s) requests that use upgrade to switch - like web sockets (ws and wss).  This was never a consideration for he mitm proxy (I guess since the package specifies version 3.2 and up - so it never worked).  This matters for me so we re include it per the recommendation from the package maintainers [ https://github.com/websockets/ws/pull/1099 ]