Update proxy.js - Problem with url.parse(req.url)
In the MITM, urls containing a simple quote are not encoded in the same way when it's http or https :

http request : the MITM encodes the simple quote and replace it with %27.
https request : the MITM does not encode the simple quote.

The problem is url.parse(req.url) used in the "parseHostAndPort" function when it is a http request. parse() replace the simple quote with %27.

I have rewritten the function "parseHostAndPort" so that it no longer uses the "parse()" method to get host and port. I use the regular expression instead to get them.