Added protocol specification when connecting to the server
In the previous implementation, when a client sends a proxy specification with the sec-websocket-protocol protocol, the proxy would disregard it and establish a connection to the server without this specification. This could lead to an incorrect connection and ultimately result in data not being sent. In my commit, I've modified the code to include the utilized protocols in the options and then pass them to the WebSocket constructor, ensuring the protocol specification is honored during the connection process